Add note about setuptools branch
Similar to the already existing note on the Bazel branch.

I pushed this to make my findings on setuptools-powered nanobind extension builds
available to interested users. This is work in progress, but builds already on my
machine, and follows the custom builds user guide in `nb_combined.cpp` on Linux.